backend/requirements.txt
sqlmodel>=0.0.8
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
email-validator>=2.0.0
requests>=2.31.0 
httpx>=0.27.0